[SYSINFRA-41] reflow/trace: reduce mutex contention
Summary:
This diff modifies the implementation of LocalTracer to reduce contention on the
shared mutex. We discovered that this was one of the root causes of the bug
described in [SYSINFRA-41](https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-41).

For more details see:
- [debugging doc](https://docs.google.com/document/d/11yrcaR3QEzAodPBMiQLCwsXvWVcybz7x8lkPPyvf9gU/edit#heading=h.3kv2176wcrc4)
- [post-mortem doc](https://docs.google.com/document/d/1sWsh9Cp-m_exgX3imRfcMoIv3H82f2oSmevVt1JJp6g)

Test Plan:
New and existing unit tests pass. Comparison of `TestLocalTracerContention`
before and after making this change:

```
Concurrency   Before (ms)   After (ms)
10            10            0.408
100           213           0.532
1,000         12146         14
10,000        -             120
100,000       -             850
```

To ensure that this change improves performance for "real" programs:
- ran a sample program which simulates high concurrent use of the
  localtracer with profiling enabled
- compared the mutex contention profile with and without the fix and
  observed a large improvement

To ensure that the new trace file flush logic works with pi-evaluator:
- deployed this diff to dev-blue-eks stable evaluator
- ran a sample analysis and verified that the trace was produced and
  matched the trace for the same sample analysis before this change
  (run ID: [252843](https://dev-pipeline.lims.grail.com/pipeline-ui/samples/P023JHG/252843))

Reviewers: ghorrell, pbohannon, smahadevan

Reviewed By: ghorrell

Subscribers: spolakh, dnicolaou

Differential Revision: https://phabricator.grailbio.com/D56703

fbshipit-source-id: acbcb3e