[SYSINFRA-217] reflow: reduce number of DescribeVolumes calls
Summary:
When a large number of instances are launched simultaneously, we've
observed that we frequently hit EC2 API request limits. This causes
requests to fail due to throttling and in some cases may even cause a
run to fail.

We've identified the APIs that are used the most in these scenarios, see
[SYSINFRA-181](https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-181)
for details. This diff reduces the number of `DescribeVolumes` calls,
which was one of the highest trafficked APIs.

When a new reflowlet is created, it starts up a "volume watcher" that
checks the EBS volume every minute and resizes it if necessary. Each of
these checks results in a `DescribeVolumes` call to get the current size
of the volume, however this is redundant because the volume size will
only change if reflowlet resized it. This diff addresses this by only
calling `DescribeVolumes` to get the size after a resizing operation.

Test Plan:
This change was tested by comparing the number of `RequestLimitExceeded`
errors before and after this change when running 900 simultaneous scale
test runs in `dev-blue-eks` (300 on `evaluator-canary` and 600 on
`evaluator-scaled-canary`).

In the baseline, we saw 5260 `DescribeVolumes` requests fail with
`RequestLimitExceeded`; with the fix it was reduced to 206. Full test
results are included below.

​---

We can see the number of `RequestLimitExceeded` errors by querying
CloudTrail:
```
SELECT eventname,
         errorcode,
         eventsource,
         awsregion,
         COUNT(*) count
FROM cloudtrail_logs_grail_msk_cloudtrail
WHERE errorcode = 'Client.RequestLimitExceeded'
        AND eventtime > '2021-04-08T15:38:00Z'
        AND eventtime < '2021-04-08T16:25:00Z'
GROUP BY  errorcode,awsregion, eventsource, eventname
ORDER BY  count desc
```

**Baseline**:

{F1369959}

Query interval: 2021-04-08T15:38:00Z to 2021-04-08T16:25:00Z

| eventname                    | errorcode                   | eventsource       | awsregion | count |
|------------------------------|-----------------------------|-------------------|-----------|-------|
| DescribeInstances            | Client.RequestLimitExceeded | ec2.amazonaws.com | us-west-2 | 7858  |
| **DescribeVolumes**          | Client.RequestLimitExceeded | ec2.amazonaws.com | us-west-2 | 5260  |
| DescribeSpotInstanceRequests | Client.RequestLimitExceeded | ec2.amazonaws.com | us-west-2 | 2022  |
| RunInstances                 | Client.RequestLimitExceeded | ec2.amazonaws.com | us-west-2 | 1076  |
| ...                          | ...                         | ...               | ..        | ...   |

**With fixes**:

{F1369960}

Query interval: 2021-04-10T22:00:00Z to 2021-04-10T22:40:00Z

| eventname                    | errorcode                   | eventsource       | awsregion | count |
|------------------------------|-----------------------------|-------------------|-----------|-------|
| RunInstances                 | Client.RequestLimitExceeded | ec2.amazonaws.com | us-west-2 | 1178  |
| DescribeInstances            | Client.RequestLimitExceeded | ec2.amazonaws.com | us-west-2 | 269   |
| **DescribeVolumes**          | Client.RequestLimitExceeded | ec2.amazonaws.com | us-west-2 | 206   |
| DescribeSpotInstanceRequests | Client.RequestLimitExceeded | ec2.amazonaws.com | us-west-2 | 203   |
| ...                          | ...                         | ...               | ..        | ...   |

Reviewers: smahadevan, ghorrell

Reviewed By: smahadevan

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D57298

fbshipit-source-id: 96e1703