reflow/ec2cluster: Introduce a SpotProber which caches results of probing per instance type.
Summary:
Previously, reflow used to probe whether we could run instances of a certain type
for each instance that we are trying to bring up.  This is wasteful since we
could be probing for the same instance type simultaneously.  The result of one
probing attempt is also sufficient to safely start many instances of that type
(based on the depth of probing).

This revision introduces a `spotProber` which centrally probes for each instance type.
It caches the result of probing and re-probes if the cached result has been "used" up
or if the ttl expires.

As a result, we should see a significant reduction in the number of `RunInstances`
AWS API calls (addresses SYSINFRA-219) that we make during a scale test
(and otherwise in production).

Test Plan:
Unit tests cover the functionality of spotProber.

Ran the scale tests and compared the number of `RunInstances` calls and errors without and with this change.

Athena Query used:
```
WITH temp_results AS (SELECT
    CASE
      WHEN (eventtime > '2021-04-08T15:30:00Z' AND eventtime < '2021-04-08T16:30:00Z') THEN 'scale_test_1'
      WHEN (eventtime > '2021-04-13T22:00:00Z' AND eventtime < '2021-04-13T23:00:00Z') THEN 'scale_test_3'
      ELSE 'other'
    END AS period, useragent, eventname, errorcode
  FROM cloudtrail_logs_grail_msk_cloudtrail
  WHERE awsregion = 'us-west-2' AND
  eventname IN ('RunInstances') AND
  (
    (eventtime > '2021-04-08T15:30:00Z' AND eventtime < '2021-04-08T16:30:00Z')
   OR
    (eventtime > '2021-04-13T22:00:00Z' AND eventtime < '2021-04-13T23:00:00Z')
  )
)
SELECT period, useragent, eventname, errorcode, COUNT(*) count FROM temp_results
GROUP BY period, useragent, eventname, errorcode
ORDER BY period, useragent, eventname, errorcode, count DESC
```

Results below show the API calls (and error codes where applicable) without this change (ie, `scale_test_1`) and with this change (ie, `scale_test_3`).
This change reduces the number of calls made that end up with a `DryRunOperation` error code, which can be seen below.
We can also see a significant reduction in the number `RequestLimitExceeded` errors.

|period	|useragent	|eventname	|errorcode	|count|
|------------------------------|-----------------------------|-------------------|-----------|-------|
| scale_test_1	| aws-sdk-go/1.29.24 (go1.13.15; linux; amd64)	| RunInstances	| Client.DryRunOperation	| 1218 |
| scale_test_1	| aws-sdk-go/1.29.24 (go1.13.15; linux; amd64)	| RunInstances	| Client.RequestLimitExceeded	| 1162 |
| scale_test_1	| spot.amazonaws.com	| RunInstances		| | 1217 |
| scale_test_3	| aws-sdk-go/1.29.24 (go1.13.15; linux; amd64)	| RunInstances	| Client.DryRunOperation	| 290 |
| scale_test_3	| aws-sdk-go/1.29.24 (go1.13.15; linux; amd64)	| RunInstances	| Client.RequestLimitExceeded	| 77 |
| scale_test_3	| spot.amazonaws.com	| RunInstances		| | 582 |

While we see a reduction in the calls, it still seems high compared to what we expect, but we aren't able to disambiguate calls from different evaluators (eg: stable vs canary) from these logs.  So the following are the logs from the specific stable evaluator which ran at the time and it shows that we probed only a few times for bringing up many more EC2 instances:
```
2021/04/16 06:12:10 ec2cluster: spot probing c5.9xlarge (depth=20) (attempt 0/5)
2021/04/16 06:12:11 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:12:11 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:13:56 ec2cluster: spot probing c5.9xlarge (depth=20) (attempt 0/5)
2021/04/16 06:13:58 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:14:08 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:14:17 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:14:19 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:14:29 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:16:06 ec2cluster: spot probing c5.9xlarge (depth=20) (attempt 0/5)
2021/04/16 06:16:06 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:16:07 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:16:16 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:16:17 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:17:13 ec2cluster: spot probing c5.9xlarge (depth=20) (attempt 0/5)
2021/04/16 06:17:13 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:17:48 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:18:14 ec2cluster: spot probing c5.9xlarge (depth=20) (attempt 0/5)
2021/04/16 06:18:15 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:18:16 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:18:47 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:19:04 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:19:20 ec2cluster: spot probing c5.9xlarge (depth=20) (attempt 0/5)
2021/04/16 06:19:20 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
2021/04/16 06:19:37 ec2cluster: generating ec2 spot instance request for instance type c5.9xlarge
```

The above logs were obtained using the following `Logs Insights` query (over time period `2021-04-16 06:10:00` to `2021-04-16 08:00:00`)
```
filter @logStream IN [
  'evaluator-stable-65568f8d8b-r5wj5_default_evaluator-75f32312fbc06677851d6a414db441d34fe678db9f336649603a69aa8491c4d0'
]
 | filter log like /(?i)(spot probing|generating ec2 spot instance)/
 | fields @timestamp, log
 | sort @timestamp
```

Reviewers: pboyapalli, ghorrell

Reviewed By: pboyapalli

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D57448

fbshipit-source-id: 9fe4f7e