reflow/tool: Remove the ability to pass alloc as a run flag
Summary:
In non-scheduler mode, the main evaluation of a reflow program runs
in a particular alloc and other work stealers (if enabled) will perform
other tasks.

This is no longer relevant, since we only use scheduler mode, so there
is no need for the ability to pass an alloc by command-line.

Cleaning up as per SYSINFRA-274

Test Plan: Unit tests

Reviewers: pboyapalli

Reviewed By: pboyapalli

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D58263

fbshipit-source-id: b176b08