[SYSINFRA-271] reflow/ec2cluster: retry DescribeSpotInstanceRequests calls; add logging
Summary:
In some cases, calls to `DescribeSpotInstanceRequests` return a response
indicating the spot request was fulfilled, but do not contain an instance ID.
This is an undocumented and unexpected behavior; the [AWS docs state:](https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeSpotInstanceRequests.html)

> If the status of the Spot Instance is fulfilled, the instance ID appears in
> the response and contains the identifier of the instance.

We suspect this is happening due to some kind of data propagation delay within
AWS and this diff tries to address it by retrying the request with a backoff.

Also, some logging changes are included to provide more info when we see this
issue.

Test Plan: Unit tests.

Reviewers: smahadevan

Reviewed By: smahadevan

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D58561

fbshipit-source-id: 5311777