[SYSINFRA-307] reflow/bootstrap: Refactor flags parsing and override session provider for bootstrap
Summary:
Move flag definitions and parsing code for bootstrap binaries to one place
and use them in both bootstrap binaries. This also adds the ability to
override config parameters through command-line to the bootstrap binaries.

Also, switch the grail bootstrap server to use the `ec2metadata`
session provider because the old `grailawssession` provider requires
a reflow version to be set (since D57695), but it does not make sense
to set it for the bootstrap server.

Test Plan:
Both bootstrap binaries were tested by making sure that they start.
If we see a log message that it's waiting for an image, then its good.

`grailreflowbootstrap` was tested as follows:
- First build a bootstrap binary:
```
> cd go/src/grail.com/cmd/reflowbootstrap/ && go build
```
- Use the current released reflow's config to start the bootstrap
```
> reflow config -marshal > /tmp/reflowconfig
> ./reflowbootstrap -config /tmp/reflowconfig
2021/05/06 06:31:19 no providers for type *infra.ReflowVersion
```
The bootstrap fails to start, which is the bug identified in SYSINFRA-307.
However, now we can override the `session` provider and make it work:
```
> ./reflowbootstrap -config /tmp/reflowconfig -session ec2metadata
2021/05/06 06:33:52 bootstrap server running, waiting (10m0s) for image...
```

The open source `reflowbootstrap` was tested as follows:
- First build an open source `reflow` binary:
```
> cd go/src/github.com/grailbio/reflow/cmd/buildreflow && go build && ./buildreflow -o /tmp/ossreflow
```
- We need credentials to run it, so get them, export and run the reflow to marshal the config:
```
> grail-aws eng/dev
(export the key id and secret)
> /tmp/ossreflow config -marshal > /tmp/ossconfig
```
- Build and run the bootstrap binary (and the config) make sure it starts:
```
> cd go/src/github.com/grailbio/reflow/cmd/reflowbootstrap && go build
> ./reflowbootstrap -config /tmp/ossconfig
2021/05/06 06:25:14 bootstrap server running, waiting (10m0s) for image...
```
- Try the same with flag override (in this case, use `ec2metadata` session provider)
```
./reflowbootstrap -config /tmp/ossconfig -session ec2metadata
2021/05/06 06:25:19 bootstrap server running, waiting (10m0s) for image...
```

Reviewers: ghorrell

Reviewed By: ghorrell

Subscribers: dnicolaou, pboyapalli

Differential Revision: https://phabricator.grailbio.com/D58622

fbshipit-source-id: 8212985