[SYSINFRA-358] reflow/eval: Apply concurrency limit for fileset operations only where it matters
Summary:
When dealing with relatively small filesets, do not apply concurrency limits.
The concurrency limit is in place to reduce memory usage when marshaling/unmarshaling
large cached filesets and need only be applied if we are dealing with large ones.

Also, move the tracing of cache lookups to be at the batch level instead of at each flow level.

Test Plan:
Unit and integration tests.

Ran a large bundle with this change and [[ https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-363 | confirmed that reflow cache lookups are much faster ]].

Reviewers: pboyapalli

Reviewed By: pboyapalli

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D59052

fbshipit-source-id: 4b3f516