[SYSINFRA-271] reflow/ec2cluster: increase spot request ttl and adjust ec2SpotRequestStatus retry strategy
Summary:
We observed in the evaluator logs that the reason we see "fulfilled" spot
requests without an instance ID is due to the spot request expiring before
getting fulfilled.

This diff increases the TTL of the spot requests from 1 minute to 2 minutes.
Furthermore, we adjust the retry strategy used in `ec2SpotRequestStatus` to
avoid unnecessary retries in cases were the request was canceled or closed.

Test Plan: Unit tests.

Reviewers: smahadevan

Reviewed By: smahadevan

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D58901

fbshipit-source-id: 5bfcbb8