[SYSINFRA-158] reflow: Execs return a remote log reference and we avoid copying them to the repo.
Summary:
Previously, the standard output and error logs were being copied over to the shared reflow
repository (S3).  But we also stream them to cloudwatch and hence resulting in a redundant copy.

Additionally, when dealing with a large number of execs in a run which are all writing out these logs,
we end up using a lot of memory when concurrently streaming them to S3, especially if the logs are big.
(as revealed from the benchmark tests run in [[ https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-316?focusedCommentId=55403&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-55403 | SYSINFRA-316 ]])

This revision instead makes each `Exec` return a remote logs reference and stores only
the reference in the shared repository.  (The reference is a tiny JSON object)

Additionally, we also print the reference to the `stdout` and `stderr` remote location for completed execs.

Test Plan:
Unit tests and Integration test.

Ran a sample program and confirmed that only the remote log reference is stored (and displayed).
```
> /tmp/linux/amd64/reflowswami -log debug -cache off run reflow/remote_test.rf
2021/05/13 20:12:51 reflow version reflowswami (GRAIL, go1.13.15)
...
2021/05/13 20:14:06  ->  remote_test.Main 15402594 exec   exec ..6922e62422db58813fccac0ecc1e59dfab0a59 echo "Hi!" > {{out}}
...
2021/05/13 20:14:13  <-  remote_test.Main 15402594 ok     exec 0s 4B
2021/05/13 20:14:13 remote_test.Main 15402594 /grail/src/grail/reflow/remote_test.rf:5:16:
	resources: {mem:500.0MiB cpu:1 disk:0B}
	sha256:bf3431c034417632178554e1c395c2421fabc3c18a7ac9a4ed5dc6e14e152f5d
	sha256:f89178281d2dd8859408beac2e1177d0caf58fa4a09d91afacb3ce3b7b07329f
	sha256:154025948b0ee12e3e279b42f8c25dc4dfce1809a33ae001745e7218a29cdfba
	ec2-54-190-65-11.us-west-2.compute.amazonaws.com:9000/1608617efe545c97/e89e799ac6ce40184447af4b8199e4cf2bc9b7fbb6593f7abda842a8784c99cf
	index.docker.io/library/ubuntu@sha256:89fd38d069a9a369525ade2bfb6922e62422db58813fccac0ecc1e59dfab0a59
	command:
	    echo "Hi!" > {{out}}
	where:
	stdout location: (cloudwatch) reflow /1608617efe545c97/e89e799ac6ce40184447af4b8199e4cf2bc9b7fbb6593f7abda842a8784c99cf/stdout
	stderr location: (cloudwatch) reflow /1608617efe545c97/e89e799ac6ce40184447af4b8199e4cf2bc9b7fbb6593f7abda842a8784c99cf/stderr
	result:
```
The `logs` tool shows that only the reference was stored:
```
> reflow logs e89e799ac6ce40184447af4b8199e4cf2bc9b7fbb6593f7abda842a8784c99cf | jq
{
  "Type": "cloudwatch",
  "LogGroupName": "reflow",
  "LogStreamName": "/1608617efe545c97/e89e799ac6ce40184447af4b8199e4cf2bc9b7fbb6593f7abda842a8784c99cf/stderr"
}
```

[SYSINFRA-316]: Ran a [[ https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-316?focusedCommentId=55451&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-55451 | validation scale test ]] confirming that earlier memory usage spikes (seen on the same test setup) are no longer seen with this fix.

Reviewers: pboyapalli, ghorrell

Reviewed By: pboyapalli

Subscribers: dnicolaou, pbohannon

Differential Revision: https://phabricator.grailbio.com/D59066

fbshipit-source-id: 0fb3fea