[SYSINFRA-321] reflow/localtrace: remove smallest testcase of TestLocalTracerContention to reduce flakiness
Summary:
TestLocalTracerContention runs a large number of concurrent traces and measures
the elapsed time to ensure that a linear increase in concurrency only leads to a
linear increase in execution time.

The smallest testcase of TestLocalTracerContention is more likely to fail than
the others because its expected duration is in order of tens of milliseconds and
sees more variance across test executions.

We can safely remove this testcase because the other two larger test cases cover
the same code and still verify that execution time increases linearly.

Test Plan: Unit tests.

Reviewers: smahadevan

Reviewed By: smahadevan

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D59068

fbshipit-source-id: 782b005