[SYSINFRA-254] reflow/ec2cluster: Fix cheapest computation
Summary:
Cheapest should be computed based on the instance types that are applicable
as per the config and not across all possible instance types.
Also, in `getInstanceAllocations` we short-circuit if we don't have a budget to work with.

Additional fixes ensure that the cost specified in the config is picked up properly.

Test Plan:
Unit tests.

Ran a sample program with insufficient budget and got the following logs:
```
2021/05/14 02:27:05 ec2cluster: resource requirements are unsatisfiable by current instance selection: {mem:1.0GiB cpu:1 disk:1.0GiB}
```

Custom config changes also work correctly.
Without the change:
```
> grep usd /tmp/config
      maxhourlycostusd: 0.3
> reflow -config /tmp/config config -marshal | grep usd
      maxhourlycostusd: 250
```

With the change:
```
> grep usd /tmp/config
      maxhourlycostusd: 0.3
> reflow -config /tmp/config config -marshal | grep usd
      maxhourlycostusd: 0.3
```

Reviewers: ghorrell

Reviewed By: ghorrell

Subscribers: dnicolaou, pboyapalli

Differential Revision: https://phabricator.grailbio.com/D59082

fbshipit-source-id: c8fc3d6