[SYSINFRA-274] reflow/eval: Remove Flow.ExecId
Summary:
The field `Flow.ExecId` was primarily being set for the purpose of unit testing.
In unit tests, we need the tests to be able to "rendezvous" every `Exec` placed
using `Executor.Put` so that we can set its result/error to make the test continue.

This field was also being set differently for scheduler and non-scheduler mode
and since the non-scheduler mode is being deprecated, there is now only one way that
this field was being set.

With this change however, we remove the need for this field altogether by
using a different field for the "rendezvous" mechanism.

Test Plan: Unit tests

Reviewers: ghorrell

Reviewed By: ghorrell

Subscribers: dnicolaou, pboyapalli

Differential Revision: https://phabricator.grailbio.com/D59169

fbshipit-source-id: 0c70362