[SYSINFRA-340] reflow/taskdb: Add a KeepIDAliveAndEnd method
Summary:
Adding a utility method to use for keeping a taskdb row alive and then updating
the end time (using a different context), once the original context (to keep it alive) is cancelled.

This will be used for keeping relevant taskdb rows alive for computing costs.

Test Plan: Unit tests.

Reviewers: pboyapalli

Reviewed By: pboyapalli

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D59422

fbshipit-source-id: 524d3bd