[SYSINFRA-354] reflow/pool: Create taskDB row for allocs started in a pool
Summary:
Create a row in taskdb of type `alloc` whenever an alloc is created in a pool.
This also maintains a keepalive on the row as long as the alloc is kept alive.
When the alloc is freed, the row's end time is updated.

The alloc row contains a field referring to the pool in which it was created
and the amount of resources allocated for the alloc.

Test Plan: Ran the integration tests and [[ https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-364 | confirmed ]] the creation of `alloc` rows.

Reviewers: pboyapalli

Reviewed By: pboyapalli

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D59651

fbshipit-source-id: 8df191f