[SYSINFRA-354] reflow/scheduler: Scheduler should free unusable allocs
Summary:
Scheduler always frees allocs that are unusable which could be due to many reasons:
- The scheduler's context was cancelled (ie, we are done with evaluation and shutting down everything)
- The alloc's context was cancelled.  This can happen if we no longer want to use the alloc (`MaxAllocIdleTime` passed), or if we couldn't keepalive.
- A keepalive error which is not `context` related.  This is typically due to inability to reach the alloc (eg: network issues, instance down)

In all of the above cases, we (try to) explicitly free the unusable alloc.

Test Plan:
Unit and integration tests.

The changes to `side_effect_test` is necessary due to the effect of this change (ie, allocs are freed when reflow exits).
Previously, the `side_effect_test` which needed the cloudwatch stream name, would use
`reflow list` to get the list of live allocs and figure out the stream name from the exec URI.
With this change, we instead parse the remote log location from the exec output.

Reviewers: pboyapalli

Reviewed By: pboyapalli

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D59539

fbshipit-source-id: dae15e4