[SYSINFRA-385] base integration test for pieval
Summary:
This is a very basic integration test for pieval that sets up a fully unmocked `pipeline.Evaluator` object with a functional reflow localcluster and hooks it into the local database routed through locally running director code.

Putting this out for any feedback and am working on s3 mock setup for it to be able to "fetch" and run bundles.

Test Plan: this diff only adds a new test

Reviewers: pbohannon, awissmann

Reviewed By: pbohannon

Subscribers: ghorrell, smahadevan, dnicolaou, pboyapalli

Differential Revision: https://phabricator.grailbio.com/D60122

fbshipit-source-id: 892e710