[SYSINFRA-472] reflow/sched_scale_test: fix data race in unit test
Summary:
The race detection tool indicates a data race in the `testSchedScale` code. The
problem is that `allocator.start()` and `submitter.start()` are invoked
asynchronously in separate goroutines. The race occurs on the `a.m` field, that
is the allocator's manager. The `a.m` field is written to by `allocator.start()`
goroutine. Since we also start the task submitter, it will submit tasks to the
scheduler, which will then call `Allocate` on the allocator, which reads `a.m`
[here](https://sg.eng.grail.com/grail/grail@reflow1.8.0/-/blob/go/src/github.com/grailbio/reflow/sched/scale/scheduler_scale_test.go#L376).

Test Plan:
Unit tests. Also, ran the sched scale tests with race detection enabled:
```
> cd go/src/github.com/grailbio/reflow/sched/scale; go test -race .

ok      github.com/grailbio/reflow/sched/scale  21.456s
```
Before this fix, the test failed with a data race warning.

Reviewers: smahadevan

Reviewed By: smahadevan

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D60738

fbshipit-source-id: 4277086