[SYSINFRA-539] reflow/eval: Remove TaskDB from reflow eval
Summary:
Previously, the `TaskDB` used to record the run was initialized on a per-`Runner` basis.
This is fine for single-run contexts (eg: `reflow run`), but in multiple-run contexts,
this causes problems because (as of now) `TaskDB` inherits the dynamodb table name from
the `Assoc`, which can change for each run.

This revision does the following:
- removes `TaskDB` from reflow `eval` and ensures that the `TaskDB` from the `scheduler` is used instead.
- Wherever `eval` needs a `TaskDB`, it is set on `sched.Task` after the task is processed by the `scheduler`.
- Includes changes to make the pipeline evaluator integration test pass.
- removes some unused/unnecessary code.

Test Plan:
Unit and integration tests.

For additional tests see [[ https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-591 | SYSINFRA-591 ]]

Reviewers: pboyapalli, spolakh, ghorrell, awissmann

Reviewed By: pboyapalli

Subscribers: pbohannon, dnicolaou

Differential Revision: https://phabricator.grailbio.com/D62312

fbshipit-source-id: 6c9442f