[SYSINFRA-560] reflow/syntax: use correct type for computing switch value digest
Summary: Use correct type when computing the digest of the value on which we are switching.  When computing a digest when deferring evaluation, we include the digest of that value.  The previous code was wrongly using the whole expression type to compute the digest.

Test Plan:
Run added unit test.

Run example program that previously panicked:
```
Main := switch #T(delay(0)) {
case #T(_):
    0
}
```

See that it doesn't panic.

Reviewers: smahadevan

Reviewed By: smahadevan

Subscribers: smahadevan, dnicolaou, pboyapalli

Differential Revision: https://phabricator.grailbio.com/D62185

fbshipit-source-id: 12cc83b