[SYSINFRA-574] pipeline/rds: fixes to increase db-reader role ttl
Summary:
This diff is a follow up to D62608 and fixes an oversight from that diff: we
need to update the `max_session_duration` value of `reader_role` in the
terraform config for the updated TTL to be allowed.

Without this, we were seeing this error:
```
ValidationError: The requested DurationSeconds exceeds the MaxSessionDuration set for this role.
	status code: 400, request id: 208130ba-469d-4601-b126-f5ded356e83a
```

Furthermore, in D62608, I didn't realize that there were separte ticket configs
for `users` and `engusers`, this diff also updates the configs in `engusers` to
match what was done in D62608.

Test Plan:
Ran `bazel run :terraform init && bazel run :terraform plan` in:

$GRAIL/ops/terraform/pipeline/grail-sysinfra-eng/rds/access
{P2063}

$GRAIL/ops/terraform/pipeline/msk/rds/access
{P2064}

Reviewers: jirgon, pbohannon, tcho, O9 ticket-server

Reviewed By: jirgon, tcho, O9 ticket-server

Differential Revision: https://phabricator.grailbio.com/D62995

fbshipit-source-id: 79d31b7