[SYSINFRA-570] reflow/dynamodb: Optimize predictor queries using new indices
Summary:
This revision has the following changes:
- add support for creating new indices with a sort key (sorted by `EndTime` column)
  on columns `ImgCmdId` and `Ident` (both used by `predictor`).
- the new indices project only the columns needed by the predictor, and hence we use
  a different query for these indices (while the old indices will continue to use the existing query).
- add support to use either the new or the older indices on these columns by detecting
  which ones exist when a query intends to use them, and also use the appropriate query.

`Predictor` currently fetches upto `1000` rows based on the current index and then sorts them
by `EndTime` before using the results.  The new indices will optimize the query and therefore
we can fetch fewer (`250`) rows from `TaskDB`, while maintaining (if not improving) the quality
of prediction as well as reduce the memory overhead incurred in processing rows.

We will need to approach this as a two-step process.  That is:
- we first release a reflow version and make sure all taskdbs are updated with the new indices.
- and in a subsequent release delete the older indices so that they don't continue to accrue costs.

For the open source `reflow` distribution, support for creating `TaskDB`
is added in the same release which includes this change, so we do not have to worry
about deleting older indices there.

Test Plan:
Unit and integration tests will ensure functionality with current indices.
For more tests, see [[ https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-696 | SYSINFRA-696 ]]

Reviewers: pboyapalli, ghorrell, jirgon

Reviewed By: pboyapalli

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D66904

fbshipit-source-id: 0bb0860