[SYSINFRA-700] reflow/tool: Avoid holding onto the Bundle in memory
Summary:
The bundle in `tool.Eval` is populated so that we can upload the bundle to the repo,
and refer it from the taskdb row pertaining to the reflow run.

This change accomplishes the same without holding onto the bundle in `tool.Eval`
throughout the lifetime of the run.

As an example, the raw source bytes of a recent pipeline (which caused evaluator OOMs)
is around `436KiB` and since 300 runs were running in parallel, we used up `~120MiB` of memory
unnecessarily.

Though minor, this is a useful improvement.

Test Plan:
Unit and integration tests.

Also ran a reflow program and confirmed that the bundle got uploaded.  For example, before this change:
```
> reflow -log debug  -cache off run reflow/remote_test.rf
2021/09/16 20:53:36 reflow version reflow1.10.0 (GRAIL, go1.13.15)
...
2021/09/16 20:53:38 created bundle sha256:6329d8f91523a4fdf6959764d3c6fd85748d1a9e609f415b41e6353ab24fbb85 with args: []
...

> reflow cat sha256:6329d8f91523a4fdf6959764d3c6fd85748d1a9e609f415b41e6353ab24fbb85 | gunzip
// Simple test to be run on EC2 to check if remote reflow runs are working.
// This test is meant to be run with cache off.

@requires(cpu := 1)
val Main = exec(image := "619867110810.dkr.ecr.us-west-2.amazonaws.com/ubuntu:16.04-1507842750") (out file) {"
	echo "Hi!" > {{out}}
"}
```
With this change:
```
> /tmp/linux/amd64/reflowswami -log debug  -cache off run reflow/remote_test.rf
...
2021/09/16 20:56:27 reflow version reflowswami (GRAIL, go1.13.15)
...
2021/09/16 20:56:29 created bundle sha256:6329d8f91523a4fdf6959764d3c6fd85748d1a9e609f415b41e6353ab24fbb85 with args: []

> reflow cat sha256:6329d8f91523a4fdf6959764d3c6fd85748d1a9e609f415b41e6353ab24fbb85 | gunzip
// Simple test to be run on EC2 to check if remote reflow runs are working.
// This test is meant to be run with cache off.

@requires(cpu := 1)
val Main = exec(image := "619867110810.dkr.ecr.us-west-2.amazonaws.com/ubuntu:16.04-1507842750") (out file) {"
	echo "Hi!" > {{out}}
"}
```

Reviewers: pboyapalli

Reviewed By: pboyapalli

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D64936

fbshipit-source-id: 220ca7c